split FullGamePacketHandler into two handlers
