+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/network/BedrockClientSession.java
@@ -125,9 +125,6 @@ public class BedrockClientSession extends Thread {
     }
 
     private void processIncomingPackets() {
-        for (BaseBedrockPacketHandler handler : this.handlers) {
-
-        }
         while (this.queuedIncomingPackets.peek() != null) {
             BaseBedrockPacket packet = this.queuedIncomingPackets.poll();
             for (BaseBedrockPacketHandler handler : this.handlers) {
