+++ pizzaserver-server/src/main/java/io/github/willqi/pizzaserver/server/network/handlers/ChunkBlockPacketHandler.java
@@ -1,0 +1,28 @@
+package io.github.willqi.pizzaserver.server.network.handlers;
+
+import io.github.willqi.pizzaserver.api.player.Player;
+import io.github.willqi.pizzaserver.server.network.BaseBedrockPacketHandler;
+import io.github.willqi.pizzaserver.server.network.protocol.packets.*;
+
+/**
+ * Handles any packets regarding chunk/block interactions
+ */
+public class ChunkBlockPacketHandler extends BaseBedrockPacketHandler {
+
+    private final Player player;
+
+
+    public ChunkBlockPacketHandler(Player player) {
+        this.player = player;
+    }
+
+    @Override
+    public void onPacket(RequestChunkRadiusPacket packet) {
+        this.player.setChunkRadiusRequested(packet.getChunkRadiusRequested());
+
+        ChunkRadiusUpdatedPacket chunkRadiusUpdatedPacket = new ChunkRadiusUpdatedPacket();
+        chunkRadiusUpdatedPacket.setRadius(this.player.getChunkRadius());
+        this.player.sendPacket(chunkRadiusUpdatedPacket);
+    }
+
+}
